backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
black==24.8.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.2
contourpy==1.1.1
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@65184fc057d4fab080a98564f6b60fae0b94edc4
dnspython==2.6.1
easydict==1.13
einops==0.8.1
email-validator==2.2.0
exceptiongroup==1.3.0
fairscale==0.4.13
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.57.0
fvcore==0.1.5.post20221221
google-auth==2.40.2
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iopath==0.1.9
jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.4
lit==18.1.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.1.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.3.56
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycocotools==2.0.7
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pywavelets==1.4.1
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
scikit-image==0.21.0
scipy==1.10.1
setuptools==58.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2023.7.10
timm==0.5.4
tomli==2.2.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.67.1
triton==2.0.0
typer==0.15.4
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
werkzeug==3.0.6
wget==3.2
wheel==0.45.1
yacs==0.1.8
zipp==3.20.2
